add security check for crud actions

+++ src/ComfinyMeliBlueBundle/Controller/CRUDController.php
@@ -21,26 +21,39 @@ class CRUDController extends Controller
      */
     public function  newAction(Request $request){
 
-        $post = new Post();
+        if($this->get('security.authorization_checker')->isGranted('ROLE_ADMIN'))
+        {
+            if($this->get('security.authorization_checker')->isGranted('IS_AUTHENTICATED_FULLY'))
+            {
+                $post = new Post();
 
-        $form = $this->createForm(PostType::class, $post);
+                $form = $this->createForm(PostType::class, $post);
 
-        $form->handleRequest($request);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $post = $form->getData();
-            $post->setCreatedAt(new \DateTime('now'));
-            $post->setModifiedAt(new \DateTime('now'));
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($post);
-             $em->flush();
-            $request->getSession()->getFlashBag()->add('Notification', 'Article bien enregistré.');
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $post = $form->getData();
+                    $post->setCreatedAt(new \DateTime('now'));
+                    $post->setModifiedAt(new \DateTime('now'));
+                    $em = $this->getDoctrine()->getManager();
+                    $em->persist($post);
+                    $em->flush();
+                    $request->getSession()->getFlashBag()->add('Notification', 'Article bien enregistré.');
 
-            return $this->redirectToRoute('blog_home',array("page" => 1));
+                    return $this->redirectToRoute('blog_home',array("page" => 1));
+                }
+                return $this->render('ComfinyMeliBlueBundle:CRUD:add.html.twig', array(
+                    'form' => $form->createView(),
+                ));
+
+            }
+            else
+            {
+                $this->get('session')->getFlashBag()->add('erreur', "Pour une question de sécurtité, l'accès à l'edition nécessite d'être connecté sans le système d'auto-connexion.");
+                return $this->redirectToRoute('fos_user_security_login');
+            }
         }
-        return $this->render('ComfinyMeliBlueBundle:CRUD:add.html.twig', array(
-            'form' => $form->createView(),
-        ));
+        return $this->render('errorAccessD.html.twig');
 
     }
 
@@ -51,26 +64,39 @@ class CRUDController extends Controller
      */
     public function  editAction(Request $request, $slug){
 
-        $postRepository = $this->getDoctrine()->getRepository(Post::class);
-        $post = $postRepository->findOneBy(array('slug' =>$slug));
+        if($this->get('security.authorization_checker')->isGranted('ROLE_ADMIN'))
+        {
+            if($this->get('security.authorization_checker')->isGranted('IS_AUTHENTICATED_FULLY'))
+            {
+                $postRepository = $this->getDoctrine()->getRepository(Post::class);
+                $post = $postRepository->findOneBy(array('slug' =>$slug));
 
-        $form = $this->createForm(PostType::class, $post);
+                $form = $this->createForm(PostType::class, $post);
 
-        $form->handleRequest($request);
+                $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $post = $form->getData();
-            $post->setModifiedAt(new \DateTime('now'));
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($post);
-            $em->flush();
-            $request->getSession()->getFlashBag()->add('Notification', 'L\'article a bien modifié.');
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $post = $form->getData();
+                    $post->setModifiedAt(new \DateTime('now'));
+                    $em = $this->getDoctrine()->getManager();
+                    $em->persist($post);
+                    $em->flush();
+                    $request->getSession()->getFlashBag()->add('Notification', 'L\'article a bien modifié.');
 
-            return $this->redirectToRoute('blog_home',array("page" => 1));
+                    return $this->redirectToRoute('blog_home',array("page" => 1));
+                }
+                return $this->render('ComfinyMeliBlueBundle:CRUD:add.html.twig', array(
+                    'form' => $form->createView(),
+                ));
+
+            }
+            else
+            {
+                $this->get('session')->getFlashBag()->add('erreur', "Pour une question de sécurtité, l'accès à l'edition nécessite d'être connecté sans le système d'auto-connexion.");
+                return $this->redirectToRoute('fos_user_security_login');
+            }
         }
-        return $this->render('ComfinyMeliBlueBundle:CRUD:add.html.twig', array(
-            'form' => $form->createView(),
-        ));
+        return $this->render('errorAccessD.html.twig');
 
     }
 
@@ -78,13 +104,26 @@ class CRUDController extends Controller
      * @Route("/delete/{id}",  name="post_delete")
      */
     public function  deleteAction($id){
-        $postRepository = $this->getDoctrine()->getRepository(Post::class);
-        $post = $postRepository->find($id);
+        if($this->get('security.authorization_checker')->isGranted('ROLE_ADMIN'))
+        {
+            if($this->get('security.authorization_checker')->isGranted('IS_AUTHENTICATED_FULLY'))
+            {
+                $postRepository = $this->getDoctrine()->getRepository(Post::class);
+                $post = $postRepository->find($id);
 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($post);
-        $em->flush();
-        return $this->redirectToRoute('blog_home',array("page" => 1));
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($post);
+                $em->flush();
+                return $this->redirectToRoute('blog_home',array("page" => 1));
+            }
+            else
+            {
+                $this->get('session')->getFlashBag()->add('erreur', "Pour une question de sécurtité, l'accès à l'edition nécessite d'être connecté sans le système d'auto-connexion.");
+                return $this->redirectToRoute('fos_user_security_login');
+            }
+        }
+        return $this->render('errorAccessD.html.twig');
+
 
     }
 }
